some basic changes

+++ src/main/java/com/studyportalbackend/SpringStudyBackend/security/jwt/JwtUtils.java
@@ -3,6 +3,7 @@ package com.studyportalbackend.SpringStudyBackend.security.jwt;
 import com.studyportalbackend.SpringStudyBackend.service.UserDetailsImpl;
 import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
 import jakarta.servlet.http.HttpServletRequest;
@@ -11,6 +12,7 @@ import org.springframework.stereotype.Component;
 
 import javax.crypto.SecretKey;
 import java.security.Key;
+import java.security.SignatureException;
 import java.util.Date;
 import java.util.stream.Collectors;
 
@@ -21,7 +23,7 @@ public class JwtUtils {
     private String jwtSecret;
 
     @Value("${jwt.expiration}")
-    private long jwtExpirationMs;
+    private int jwtExpirationMs;
 
     public String getJwtFormHeader(HttpServletRequest request) {
         String bearerToken = request.getHeader("Authorization");
